+++ data/pool_data.c
@@ -2,5 +2,5 @@
 #include "pool_data.h"
 
 EnemyData enemy_data[ENEMY_DATA_COUNT] = {
-    { { 10, 0, 0, 0 }, 300, 100, 0, 1 },
+    { { 10, 0, 0, 0 }, 32, 92, 0, 1 },
 };
