+++ movement/parse.h
@@ -3,8 +3,8 @@
 #include "../data/pool_data.h"
 
 // Reads in the data for enemies and obstacles
-void read_enemy(Enemy **enemies, EnemyData *data, UINT8 enemy_data_count, UINT8 *enemy_num, UINT8 latest_id);
+extern Enemy* read_enemy(EnemyData *data, UINT8 enemy_data_count, UINT8 *enemy_num, UINT8 latest_id);
 
-void clear_enemy(Enemy **enemies, UINT8 id);
+extern void clear_enemy(Enemy **enemies, UINT8 id);
 
-void empty_enemy(Enemy **enemies);
+extern void empty_enemy(Enemy **enemies);
